Store the keychain items with the actual GitHub username
This adds a new config file that we use just to store the username (so
taht we can look it up again in the future). Then it updates things so
that instead of always writing as `github_token` was the username we use
the _real_ username.

In order to get the real username to begin with, we need to ping the
`GET /user` endpoint.